fix: change type enum to integer for rating column

+++ src/database/migrations/2025_01_07_110710_create_supplier_ratings_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('supplier_id')->constrained()->onDelete('cascade');
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->decimal('rating', 3, 1);
+            $table->integer('rating');
             $table->text('review')->nullable();
             $table->timestamps();
         });
